+++ viewflow/__init__.py
@@ -13,7 +13,7 @@ from .this_object import this
 from .utils import viewprop, Icon, DEFAULT
 
 __title__ = "Django-Viewflow"
-__version__ = "2.3.0"
+__version__ = "2.3.1"
 __author__ = "Mikhail Podgurskiy"
 __license__ = "AGPL"
 __copyright__ = "Copyright 2018-2021 Mikhail Podgurskiy"